feat(player): implementa habilidade de atordoamento em área
