+++ Assets/_Game/Gameplay/Player/Violence.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da6dcf5d503d87740a958c3f4eb53510
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
